remove hero image from GamePage, update month formatting in ChartAreaInteractive, and enhance ParticipationPerDayChart with debug information and improved chart layout

+++ components/ChartAreaInteractive.tsx
@@ -122,7 +122,7 @@ export function ChartAreaInteractive() {
                                 const date = new Date(value)
                                 return date.toLocaleDateString("fr-FR", {
                                     day: "2-digit",
-                                    month: "short",
+                                    month: "numeric",
                                 })
                             }}
                         />
